requirements.txt
python-dotenv
openai
requests
feedparser
elevenlabs
moviepy
